DateRangePicker: rename from DateRange and move to regular exports
