+++ devbox/apps/index.js
@@ -14,7 +14,7 @@ export { default as Card } from './Card'
 export { default as Checkbox } from './Checkbox'
 export { default as ContextMenu } from './ContextMenu'
 export { default as DataView } from './DataView'
-export { default as DateRange } from './DateRange'
+export { default as DateRangePicker } from './DateRangePicker'
 export { default as DiscButton } from './DiscButton'
 export { default as Distribution } from './Distribution'
 export { default as DropDown } from './DropDown'
